Isolate filter operations from engine facade
apply_filter and its filter helpers now live in a dedicated module while mcp_video.engine keeps compatibility exports for apply_filter, _get_color_preset_filter, and _build_pitch_shift_filter. The extraction preserves existing filter behavior and tightens escaped numeric interpolation for filter values.

Constraint: Keep convert and timeline/composite operations untouched; apply_filter has broad coverage but a larger FFmpeg filter surface.

Rejected: Redesign filter dispatch as a registry | this PR is an extraction and sanitization pass, not a feature redesign.

Confidence: high

Scope-risk: broad

Directive: Filter string changes are security-sensitive; keep focused tests around every filter family before modifying mappings.

Tested: ruff check mcp_video/engine.py mcp_video/engine_filters.py; ruff format --check mcp_video/engine.py mcp_video/engine_filters.py; python3 -m pytest tests/test_engine_advanced.py -k 'ApplyFilter or ColorPresetFilter or FilterValidation or AudioEffects or KenBurns' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'filter or blur' tests/test_server.py -k 'filter or blur' tests/test_red_team.py -k 'apply_filter' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
